agents/sre-multi-agent/fast_agentcore/patterns/langgraph-single-agent/requirements.txt
# LangGraph agent dependencies with pinned versions
langgraph==1.0.5
langchain-aws==1.0.0
langchain-mcp-adapters==0.1.13
langgraph-checkpoint-aws==1.0.1
mcp==1.23.1
bedrock-agentcore==1.0.6
PyJWT[crypto]>=2.10.1

# SRE Agent dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0
boto3>=1.34.0
